# digital-twingo/scene-server-go/docs/swagger.json
{
    "swagger": "2.0",
    "info": {
        "description": "三维数字孪生场景设计器接口文档",
        "title": "三维数字孪生场景设计器接口文档",
        "contact": {
            "name": "北京煜邦电力技术股份有限公司",
            "url": "www.yupont.com",
            "email": "tiany@yupont.com"
        },
        "version": "v1.0.0"
    },
    "paths": {
...
        "/sceneApi/datasvr/getData": {
            "get": {
                "description": "根据dataId返回对应数据对象，实际项目中需要扩充此接口，依据不同类型返回具体的数据项目",
                "produces": [
                    "application/json"
                ],
                "tags": [
                    "数据服务接口"
                ],
                "summary": "数据服务",
                "parameters": [
                    {
                        "type": "string",
                        "description": "数据ID",
                        "name": "dataId",
                        "in": "query",
                        "required": true
                    }
                ],
                "responses": {
                    "200": {
                        "description": "OK",
                        "schema": {
                            "$ref": "#/definitions/vo.ResultVo"
                        }
                    }
                }
            }
        },
...
    },
    "definitions": {
        "vo.ResultVo": {
            "type": "object",
            "properties": {
                "code": {
                    "type": "integer"
                },
                "data": {}
            }
        }
    }
}